first steps of deprecating the old catalog engine
