+++ .changeset/tricky-poems-decide.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend': patch
+---
+
+First steps of deprecating the old catalog engine
