+++ plugins/catalog-backend/src/index.ts
@@ -23,6 +23,7 @@
 export * from './catalog';
 export * from './database';
 export * from './ingestion';
+export * from './legacy';
 export * from './search';
 export * from './service';
 export * from './util';
